Created a Zapper and a Checkpoint
The player teleports to the most recently touched checkpoint if the
touch a zapper.  The checkpoint is supposed to light up when it detects
a collision with the player, but for some reason the checkpoint doesn't
detect collisions with the player even when the player detects a
collision with the checkpoint.

Changes to the player movement scripts were in the OnTriggerEnter2D
methods, along with some class variables.  The checkpoint and zappers
are prefabs, and they should come with the appropriate tags equipped,
but if not, this is a potential source for bugs.