Added cannon object
Level1 has a man on the bottom track that functions as a placeholder cannon.
The cannon shoots a player to the right until the player hits an object with the "Wall" tag.
This persists between swapping.
To use a cannon, walk into the cannon's collider and press the jump button.
To make a new cannon, give an object the "Cannon" tag.
TODO: Add a sprite for the cannon, add a sprite that appears above a cannon when a player is within the area that shows a "Jump" icon.