A short level, with 3 major obstacles.  Also made some changes to allow the cannon to shoot in different directions.
Cannon now has a cannon controller script attatched that allows you to
specify x and y velocity.
I also reduced gravity to .8 to make things a little easier.  This does
change some jump heights so if this screws up other peoples levels I can
put a fan or something in my level that blows air upwards so it feels
more natural that this particular level has lighter gravity, or we can
switch it back and just have it be decently harder (its only really
impactful on the jumps to the top floor of the upper track as you have
to make an extra jump and swap if using base gravity.)