Level 3, created button prefab, cannons can be activated/deactivated
Level 3 is a swap while falling map. Cannon is disabled until a player jumps onto the button at the end of the bottom track.  (Buttons look suspiciously like tables)

Cannons now have bouncing text that displays when a player is within the trigger area of the cannon.

Once slight air movement is implemented, the buttons should be much easier to press.