Added 2 small "tutorial" levels, renamed levels, updated level select
Levels 1 and 2 are tutorials on swapping and using the cannon. 
Renames the levels so we have a proper ordering.
Updated level select menu to include new levels, made the buttons big squares (going to put a slightly transparent screenshot of the level as each button's background), and changed the "back" button to a "main menu" button since "back" implied it would go back to the level you were just on if you went to the level select menu from a level's pause menu.